requirements.txt
Flask==1.0.2
flask_dance==2.0.0
Flask-SQLAlchemy==2.3.2
bcrypt==3.1.6
SQLAlchemy==1.3.2
psycopg2==2.7.7
PyJWT==1.7.1
